python/pyproject.toml
[tool.poetry]
name = "predictor"
version = "0.1.0"
description = ""
authors = ["Kirill Parkhomenko <kxrxh.contact@gmail.com>"]
packages = [
    { include = "shared", from = "src" },
    { include = "cli", from = "src" },
    { include = "server", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
pandas = "^2.2.3"
catboost = "^1.2.7"
joblib = "^1.4.2"
pyyaml = "^6.0.1"
click = "^8.1.7"
rich = "^13.9.1"
scikit-learn = "^1.5.2"
uvicorn = "^0.24.0"
python-multipart = "^0.0.9"
pyarrow = "^14.0.0"
tenacity = "^8.3.0"

[tool.poetry.group.server]
optional = true

[tool.poetry.group.server.dependencies]
fastapi = { extras = ["standard"], version = "^0.115.0" }

[tool.poetry.group.cli]
optional = true

[tool.poetry.group.cli.dependencies]
click = "^8.1.7"
rich = "^13.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "cli.main:cli"
server = "server.main:main"
